requirements.txt
bs4
feedparser
httpx
python-telegram-bot
